refactor: split physics module
