+++ crates/physics/src/simulation.rs
@@ -1,0 +1,275 @@
+use crate::types::{Joint, JointParams, PhysParams, Sphere, Vec3};
+use compute::{ComputeBackend, ComputeError};
+use std::mem::size_of;
+use std::sync::Arc;
+
+pub struct SphereState {
+    pub pos: Vec3,
+}
+
+#[derive(Debug)]
+pub enum PhysicsError {
+    BackendError(ComputeError),
+    NoSpheres,
+}
+
+impl From<ComputeError> for PhysicsError {
+    fn from(err: ComputeError) -> Self {
+        PhysicsError::BackendError(err)
+    }
+}
+
+pub struct PhysicsSim {
+    pub spheres: Vec<Sphere>,
+    pub params: PhysParams,
+    pub joints: Vec<Joint>,
+    pub joint_params: JointParams,
+    backend: Arc<dyn ComputeBackend>,
+}
+
+const WORKGROUP_SIZE: u32 = 256;
+
+impl PhysicsSim {
+    #[must_use]
+    pub fn new_single_sphere(initial_height: f32) -> Self {
+        let sphere = Sphere {
+            pos: Vec3::new(0.0, initial_height, 0.0),
+            vel: Vec3::new(0.0, 0.0, 0.0),
+        };
+        let spheres = vec![sphere];
+
+        let params = PhysParams {
+            gravity: Vec3::new(0.0, -9.81, 0.0),
+            dt: 0.01,
+            force: [0.0, 0.0],
+        };
+
+        let backend = compute::default_backend();
+
+        Self {
+            spheres,
+            params,
+            joints: Vec::new(),
+            joint_params: JointParams {
+                compliance: 0.0,
+                _pad: [0.0; 3],
+            },
+            backend,
+        }
+    }
+
+    pub fn step_gpu(&mut self) -> Result<(), PhysicsError> {
+        if self.spheres.is_empty() {
+            return Ok(());
+        }
+        let sphere_bytes: Arc<[u8]> = bytemuck::cast_slice(&self.spheres).to_vec().into();
+        let sphere_buffer_view = compute::BufferView::new(
+            sphere_bytes.clone(),
+            vec![self.spheres.len()],
+            size_of::<Sphere>(),
+        );
+
+        let params_bytes: Arc<[u8]> = bytemuck::bytes_of(&self.params).to_vec().into();
+        let params_buffer_view = compute::BufferView::new(params_bytes, vec![1], size_of::<PhysParams>());
+
+        let num_spheres = self.spheres.len() as u32;
+        let workgroups_x = (num_spheres + WORKGROUP_SIZE - 1) / WORKGROUP_SIZE;
+
+        let result_buffers = self.backend.dispatch(
+            &compute::Kernel::IntegrateBodies,
+            &[sphere_buffer_view.clone(), params_buffer_view],
+            [workgroups_x, 1, 1],
+        )?;
+
+        if let Some(updated) = result_buffers.get(0) {
+            if updated.len() == self.spheres.len() * size_of::<Sphere>() {
+                let new_spheres: Vec<Sphere> = updated
+                    .chunks_exact(size_of::<Sphere>())
+                    .map(bytemuck::pod_read_unaligned)
+                    .collect();
+                self.spheres.clone_from_slice(&new_spheres);
+            }
+        }
+
+        #[repr(C)]
+        #[derive(Copy, Clone, Debug, bytemuck::Pod, bytemuck::Zeroable)]
+        struct SdfVec3 {
+            x: f32,
+            y: f32,
+            z: f32,
+        }
+        #[repr(C)]
+        #[derive(Copy, Clone, Debug, bytemuck::Pod, bytemuck::Zeroable)]
+        struct SdfBody {
+            pos: SdfVec3,
+        }
+        #[repr(C)]
+        #[derive(Copy, Clone, Debug, bytemuck::Pod, bytemuck::Zeroable)]
+        struct SdfPlane {
+            height: f32,
+        }
+        #[repr(C)]
+        #[derive(Copy, Clone, Debug, bytemuck::Pod, bytemuck::Zeroable)]
+        struct SdfContact {
+            index: u32,
+            penetration: f32,
+        }
+
+        let bodies: Vec<SdfBody> = self
+            .spheres
+            .iter()
+            .map(|s| SdfBody {
+                pos: SdfVec3 {
+                    x: s.pos.x,
+                    y: s.pos.y,
+                    z: s.pos.z,
+                },
+            })
+            .collect();
+
+        let bodies_bytes: Arc<[u8]> = bytemuck::cast_slice(&bodies).to_vec().into();
+        let bodies_view = compute::BufferView::new(bodies_bytes, vec![bodies.len()], size_of::<SdfBody>());
+
+        let plane = SdfPlane { height: 0.0 };
+        let plane_bytes: Arc<[u8]> = bytemuck::bytes_of(&plane).to_vec().into();
+        let plane_view = compute::BufferView::new(plane_bytes, vec![1], size_of::<SdfPlane>());
+
+        let placeholder: Arc<[u8]> = vec![0u8; bodies.len() * size_of::<SdfContact>()].into();
+        let contacts_view = compute::BufferView::new(placeholder, vec![bodies.len()], size_of::<SdfContact>());
+
+        let contact_buffers = self.backend.dispatch(
+            &compute::Kernel::DetectContactsSDF,
+            &[bodies_view, plane_view, contacts_view],
+            [1, 1, 1],
+        )?;
+
+        let contacts: Vec<SdfContact> = if let Some(bytes) = contact_buffers.get(0) {
+            bytes
+                .chunks_exact(size_of::<SdfContact>())
+                .map(bytemuck::pod_read_unaligned)
+                .collect()
+        } else {
+            Vec::new()
+        };
+
+        #[repr(C)]
+        #[derive(Copy, Clone, Debug, bytemuck::Pod, bytemuck::Zeroable)]
+        struct PbdContact {
+            body_index: u32,
+            normal: SdfVec3,
+            depth: f32,
+        }
+
+        let contacts_pbd: Vec<PbdContact> = contacts
+            .iter()
+            .map(|c| PbdContact {
+                body_index: c.index,
+                normal: SdfVec3 { x: 0.0, y: 1.0, z: 0.0 },
+                depth: c.penetration,
+            })
+            .collect();
+
+        let contacts_bytes: Arc<[u8]> = bytemuck::cast_slice(&contacts_pbd).to_vec().into();
+        let contacts_view = compute::BufferView::new(
+            contacts_bytes,
+            vec![contacts_pbd.len()],
+            size_of::<PbdContact>(),
+        );
+
+        let sphere_bytes: Arc<[u8]> = bytemuck::cast_slice(&self.spheres).to_vec().into();
+        let spheres_view = compute::BufferView::new(
+            sphere_bytes.clone(),
+            vec![self.spheres.len()],
+            size_of::<Sphere>(),
+        );
+        let params_placeholder: Arc<[u8]> = vec![0u8; 4].into();
+        let params_view = compute::BufferView::new(params_placeholder, vec![1], 4);
+
+        let solved = self.backend.dispatch(
+            &compute::Kernel::SolveContactsPBD,
+            &[spheres_view, contacts_view, params_view],
+            [1, 1, 1],
+        )?;
+
+        if let Some(bytes) = solved.get(0) {
+            if bytes.len() == self.spheres.len() * size_of::<Sphere>() {
+                let updated: Vec<Sphere> = bytes
+                    .chunks_exact(size_of::<Sphere>())
+                    .map(bytemuck::pod_read_unaligned)
+                    .collect();
+                self.spheres.clone_from_slice(&updated);
+            }
+        }
+
+        #[repr(C)]
+        #[derive(Copy, Clone, Debug, bytemuck::Pod, bytemuck::Zeroable)]
+        struct JointVec3 {
+            x: f32,
+            y: f32,
+            z: f32,
+        }
+        #[repr(C)]
+        #[derive(Copy, Clone, Debug, bytemuck::Pod, bytemuck::Zeroable)]
+        struct JointBody {
+            pos: JointVec3,
+        }
+
+        let joint_bodies: Vec<JointBody> = self
+            .spheres
+            .iter()
+            .map(|s| JointBody {
+                pos: JointVec3 {
+                    x: s.pos.x,
+                    y: s.pos.y,
+                    z: s.pos.z,
+                },
+            })
+            .collect();
+
+        let body_bytes: Arc<[u8]> = bytemuck::cast_slice(&joint_bodies).to_vec().into();
+        let body_view = compute::BufferView::new(body_bytes, vec![joint_bodies.len()], size_of::<JointBody>());
+
+        let joint_bytes: Arc<[u8]> = bytemuck::cast_slice(&self.joints).to_vec().into();
+        let joint_view = compute::BufferView::new(joint_bytes, vec![self.joints.len()], size_of::<Joint>());
+
+        let joint_param_bytes: Arc<[u8]> = bytemuck::bytes_of(&self.joint_params).to_vec().into();
+        let joint_param_view = compute::BufferView::new(joint_param_bytes, vec![1], size_of::<JointParams>());
+
+        let solved = self.backend.dispatch(
+            &compute::Kernel::SolveJointsPBD,
+            &[body_view, joint_view, joint_param_view],
+            [1, 1, 1],
+        )?;
+
+        if let Some(bytes) = solved.get(0) {
+            if bytes.len() == self.spheres.len() * size_of::<JointBody>() {
+                let updated: Vec<JointBody> = bytes
+                    .chunks_exact(size_of::<JointBody>())
+                    .map(bytemuck::pod_read_unaligned)
+                    .collect();
+                for (sphere, upd) in self.spheres.iter_mut().zip(updated) {
+                    sphere.pos.x = upd.pos.x;
+                    sphere.pos.y = upd.pos.y;
+                    sphere.pos.z = upd.pos.z;
+                }
+            }
+        }
+
+        Ok(())
+    }
+
+    pub fn run(&mut self, dt: f32, steps: usize) -> Result<SphereState, PhysicsError> {
+        if self.spheres.is_empty() {
+            return Err(PhysicsError::NoSpheres);
+        }
+        self.params.dt = dt;
+
+        for _ in 0..steps {
+            self.step_gpu()?;
+        }
+
+        Ok(SphereState {
+            pos: self.spheres[0].pos,
+        })
+    }
+}
